+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Added
+
+- Observers: Base pitch
+
 ## [0.1.0] - 2022/09/12
 
 Starting this changelog.